move artworkGrid to directory
