+++ src/Apps/WorksForYou/WorksForYouFeed.tsx
@@ -1,7 +1,7 @@
 import { Box, Separator, Spacer, Spinner } from "@artsy/palette"
 import { WorksForYouFeed_viewer } from "__generated__/WorksForYouFeed_viewer.graphql"
 import { ContextProps } from "Artsy"
-import ArtworkGrid from "Components/ArtworkGrid"
+import ArtworkGrid from "Components/ArtworkGrid/ArtworkGrid"
 import React, { Component } from "react"
 import ReactDOM from "react-dom"
 import styled from "styled-components"
