+++ src/Components/index.tsx
@@ -3,7 +3,9 @@ import React from "react"
 // @ts-ignore
 import Artwork, { ArtworkProps } from "./Artwork"
 // @ts-ignore
-import ArtworkGrid, { ArtworkGridContainerProps } from "./ArtworkGrid"
+import ArtworkGrid, {
+  ArtworkGridContainerProps,
+} from "./ArtworkGrid/ArtworkGrid"
 // @ts-ignore
 import BorderedPulldown, { BorderedPullDownProps } from "./BorderedPulldown"
 import Buttons from "./Buttons"
