+++ src/Styleguide/Components/ArtworkGridExample.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { graphql } from "react-relay"
 
 import { RootQueryRenderer } from "Artsy/Relay/RootQueryRenderer"
-import ArtworkGrid from "Components/ArtworkGrid"
+import ArtworkGrid from "Components/ArtworkGrid/ArtworkGrid"
 
 export function ArtworkGridExample(props: {
   artistID: string
